+++ docs/src/examples/QasTableGenerator/ClickableRow.vue
@@ -2,13 +2,7 @@
   <!-- Utilizando o list-view apenas para facilitar para recuperar os dados. -->
   <qas-list-view v-model:fields="fields" v-model:results="results" :entity="entity" use-auto-handle-on-delete :use-filter="false">
     <template #default>
-      <qas-table-generator :columns="columns" :fields="fields" :results="results" row-key="uuid" @row-click="rowClick">
-        <template #body-cell-actions="{ row }">
-          <div class="flex justify-end no-wrap q-gutter-x-sm">
-            <qas-delete :custom-id="row.uuid" entity="users" icon="sym_r_delete" />
-          </div>
-        </template>
-      </qas-table-generator>
+      <qas-table-generator :columns :fields :results row-key="uuid" @row-click="rowClick" />
     </template>
   </qas-list-view>
 </template>
@@ -30,8 +24,7 @@ export default {
     columns () {
       return [
         'isActive',
-        'name',
-        { align: 'right', name: 'actions' }
+        'name'
       ]
     }
   },